reverting fonts back
