+++ scripts/scripts.js
@@ -21,19 +21,19 @@ function clearTopBannerDismissedOnLoad() {
   }
 }
 
-// /**
-//  * load fonts.css and set a session storage flag
-//  */
-// async function loadFonts() {
-//   await loadCSS(`${window.hlx.codeBasePath}/styles/fonts.css`);
-//   try {
-//     if (!window.location.hostname.includes('localhost')) {
-//       sessionStorage.setItem('fonts-loaded', 'true');
-//     }
-//   } catch (e) {
-//     // do nothing
-//   }
-// }
+/**
+ * load fonts.css and set a session storage flag
+ */
+async function loadFonts() {
+  await loadCSS(`${window.hlx.codeBasePath}/styles/fonts.css`);
+  try {
+    if (!window.location.hostname.includes('localhost')) {
+      sessionStorage.setItem('fonts-loaded', 'true');
+    }
+  } catch (e) {
+    // do nothing
+  }
+}
 
 /**
  * Builds all synthetic blocks in a container element.
@@ -167,14 +167,14 @@ async function loadEager(doc) {
     await waitForLCP(LCP_BLOCKS);
   }
 
-  // try {
-  //   /* if desktop (proxy for fast connection) or fonts already loaded, load fonts.css */
-  //   if (window.innerWidth >= 900 || sessionStorage.getItem('fonts-loaded')) {
-  //     loadFonts();
-  //   }
-  // } catch (e) {
-  //   // do nothing
-  // }
+  try {
+    /* if desktop (proxy for fast connection) or fonts already loaded, load fonts.css */
+    if (window.innerWidth >= 900 || sessionStorage.getItem('fonts-loaded')) {
+      loadFonts();
+    }
+  } catch (e) {
+    // do nothing
+  }
 }
 
 /**
@@ -195,7 +195,7 @@ async function loadLazy(doc) {
   loadTopBanner(doc);
 
   loadCSS(`${window.hlx.codeBasePath}/styles/lazy-styles.css`);
-  // loadFonts();
+  loadFonts();
 
   sampleRUM('lazy');
   sampleRUM.observe(main.querySelectorAll('div[data-block-name]'));
